refactored code for summary filter
