+++ src/app/modules/search/components/filter-summary-container/components/filter-summary-filter/filter-summary-filter.component.html
@@ -1,4 +1,4 @@
-<span>{{filter}}</span>
-<span id="{{filter}}" class="ml-2 cross-btn" (click)="removeFilterClicked(filter)">
+<span>{{filterValue}}</span>
+<span class="ml-2 cross-btn" (click)="removeFilterClicked()">
   &#10005;
 </span>
