+++ src/app/modules/search/components/filter-summary-container/components/filter-summary-filter/filter-summary-filter.component.ts
@@ -6,15 +6,15 @@ import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
   styleUrls: ['./filter-summary-filter.component.scss']
 })
 export class FilterSummaryFilterComponent implements OnInit {
-  @Input() filter;
-  @Output() removeFilter = new EventEmitter<string>();
+  @Input() filterValue: string;
+  @Output() removeFilter = new EventEmitter();
 
   constructor() { }
 
-  ngOnInit() {}
+  ngOnInit() { }
 
-  removeFilterClicked(id: string) {
-    this.removeFilter.emit(id);
+  removeFilterClicked() {
+    this.removeFilter.emit();
   }
 
 }
