+++ src/app/modules/search/components/filter-summary-container/components/filter-summary-sort-by/filter-summary-sort-by.component.ts
@@ -24,7 +24,7 @@ export class FilterSummarySortByComponent implements OnInit {
 
   sortOrder(sortOption: SortFilter) {
     this.currentSort = sortOption;
-    this.selectedSort.emit({name: 'sort', value: sortOption.value});
+    this.selectedSort.emit({ name: 'sort', value: sortOption.value });
   }
 
 }
