+++ src/environments/environment.dev-ng-spree.ts
@@ -2,8 +2,8 @@ import { DEFAULT_CONFIG } from '../config/default/default';
 
 export const environment = {
   production: false,
-  // apiEndpoint: DEFAULT_CONFIG.prodApiEndpoint,
-  apiEndpoint: 'http://localhost:3000/',
+  apiEndpoint: DEFAULT_CONFIG.prodApiEndpoint,
+  // apiEndpoint: 'http://localhost:3000/',
   appName: DEFAULT_CONFIG.appName,
   config: DEFAULT_CONFIG
 };
